+++ src/main/java/ticketaka/mtvs3_final_backend/redis/FileUpload/domain/UploadStatus.java
@@ -1,4 +1,4 @@
-package ticketaka.mtvs3_final_backend.redis.identification.domain;
+package ticketaka.mtvs3_final_backend.redis.FileUpload.domain;
 
 public enum UploadStatus {
     // 인증 대기, 인증 완료, 인증 실패
