+++ src/main/java/ticketaka/mtvs3_final_backend/redis/FileUpload/repository/FileUploadRedisRepository.java
@@ -1,7 +1,7 @@
-package ticketaka.mtvs3_final_backend.redis.identification.repository;
+package ticketaka.mtvs3_final_backend.redis.FileUpload.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import ticketaka.mtvs3_final_backend.redis.identification.domain.FileUpload;
+import ticketaka.mtvs3_final_backend.redis.FileUpload.domain.FileUpload;
 
 import java.util.Optional;
